Added GASS driver.  Coded by Saurabh Bagchi, bagchi@uiuc.edu. Minor changes done to test/gass_xxx.c so that they print the test skip message when GASS driver is not available.
This change is the implementation of GASS within HDF5-1.3 (HDF5 with Virtual
File Layer). The GASS driver gives the facility of accessing HDF files on
remote ftp servers. To use the GASS driver, the option --with-gass=<GASS path>
shoud be specified with configure. An example of the command line used to
test the distribution was:
./configure --disable-shared --without-hdf4
--with-gass=/afs/ncsa/projects/hdf/v5/bagchi/globus/GLB/development/sparc-sun-solaris2.6_nothreads_standard_debug/include,/afs/ncsa/projects/hdf/v5/bagchi/globus/GLB/development/sparc-sun-solaris2.6_nothreads_standard_debug/lib
--disable-parallel

The user should change the path to point to his local GASS installation.

Documentation about the features of GASS and the HDF-GASS design is available
separately and till it is put up on the official web site, anyone interested
may contact me.

Test programs to read, write or append remote files have been provided in the
test directory as "gass_read.c", "gass_write.c", "gass_append.c". The test
programs have the ftp site to access #define-d at the top of the file which
the user can change accordingly.

./src/H5Epublic.
	Added new error type for file close.
./src/H5F.c
	Added hooks for the GASS driver.
./src/H5public.h
	Added header files for GASS & Globus.
./src/Makefile.in
	Added dependancy on GASS driver in LIB_SRC.
./src/hdf5.h
	Included header file for GASS driver.
./src/H5FDgass.c	[NEW]
	Routines for the GASS driver.
./src/H5FDgass.h	[NEW]
	Header file for the GASS driver.

./test/Makefile.in
	Added dependancy on the gass test routines.
./test/gass_read.c
	File to test remote file reading using GASS.
./test/gass_write.c
	File to test remote file writing using GASS.
./test/gass_append.c
	File to test remote file appending using GASS.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@1621 dab4d1a6-ed17-0410-a064-d1ae371a2980