Purpose: Configuration improvement
Problem:
Arabica has a zlib library that is a dynamic lib in /usr/ncsa/lib.
Binary generated would fail to run with a missig libz.so complain
if /usr/ncsa/lib is not in $LD_LIBRARY_PATH.  This exposed the
problem that if the binary is linked with -L$mylibpath -lxyz
where libxyz is a dynamic type, the binary can't run if $mylibpath
is not in $LD_LIBRARY_PATH (or equivalent) or as part of the
system default library paths (e.g., /usr/lib, /usr/local/lib).
This problem also caused failures during configure (when trying
to see what format to print long long) and during H5detect.
The base cause is because libtool does not "transfer" the
knowledge of -L$mylibpath to the wrapper file its mode=link
generates.

Solution:
bin/ltmain.sh:
    Changed it so that it transfers the libpaths from -Llibpaths
    to the wrapper generated.  I used the already defined variable
    finalize_shlibpath.  Not sure it is correct to use it this way.
    Need to check on this.
configure.in:
    Put in a patch to transfer information from LDFLAGS to LD_LIBRARY_PATH
    right before the AC_TRY_RUN.  The above fix for ltmain.sh does not
    work here because libtool is generated later than this point.
    There should be a cleaner way to do this.
    Removed the hardcode of NCSA_LT_LINK_EXE because it is no longer
    needed.
commence.in:
    Removed the hardcode of NCSA_LT_LINK_EXE because it is no longer
    needed.
Makefile.in:
    Another patch of LDFLAGS to LD_LIBRARY_PATH because I could not
    make it to generate H5detect in the wrapper form of libtool.
    Need to make it to use the libtool correctly. Later.

Platform tested:
arabica (solaris 2.7)
baldric (solaris 2.6)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@1975 dab4d1a6-ed17-0410-a064-d1ae371a2980