Purpose: Bug fix.
Description:
n_test must be inited to 1 when it is for tests within one machine.
Undid my previous mistake.

Platforms tested:
Eirene.
Did not do the h5committest because this is a script and if it works
in one machine, it most like will work in all Unix platform.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@6523 dab4d1a6-ed17-0410-a064-d1ae371a2980