Purpose: Feature (bug fix)
Description:
Added a new configure file specifically for the IA64 platform
in which ecc and efc are the default compilers.
Also added all the Fortran special setup in the fortran version

Platforms tested:
no h5committest since this is a IA64 platform stuff.
Tested in Titan.

Misc. update:
MANIFEST Updated.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@6565 dab4d1a6-ed17-0410-a064-d1ae371a2980