Purpose:     Code cleanup, bug fix & new feature.
Description:
    The local heaps were missed in the last changes to the metadata "dirty"
flag, making them operate incorrectly in the metadata cache.  Fixed by changing
to use "cache_info.dirty" flag instead.

    Track changes for internal metadata API, regarding marking metadata as clean
without flushing it to the file.

    Added ability to delete a local heap in the file.

    Added ability to reduce the size of a  local heap if there is a free-space
block at the end of the heap and the free-space block is over half the size of
the local heap.

    Some migration of package-specific information out of the private header
and into the code module as well.


Platforms tested:
    FreeBSD 4.8 (sleipnir) w/C++
    Linux 2.4 (burrwhite) w/FORTRAN
    Solaris 2.7 (arabica) w/FORTRAN
    IRIX64 6.5 (modi4) w/parallel & FORTRAN

    (h5committest not run due to my ongoing difficulties with C++ on burrwhite).


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@6657 dab4d1a6-ed17-0410-a064-d1ae371a2980