Purpose: Typo fix.
Description:
Found a typo in the enable-asyncio processing.  Did not get exposed
unless you ran it in an asyncio working system that uses a zlib that is
installed in a non-system area.

Solution:
Fixed the typo in CPPFLAGS.

Platforms tested:
Tested at LLNL Blue only--that is the only system that asyncio works
currently.  Pretty sure the fix won't break other systems.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@6716 dab4d1a6-ed17-0410-a064-d1ae371a2980