Purpose: New tool.
Description:
A new tool to build the HDF5 library.  This was historically the
makeh5 command.  Added here so that it can be used in machines outside
of the HDF Groups.  Having it in the source allows version specific
customization.

Platforms tested:
Did run h5committest which failed due to other source code problems.
Tested it by hand by ../hdf5/bin/buildhdf5 and passed as far as its
own functionality.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@7195 dab4d1a6-ed17-0410-a064-d1ae371a2980