Purpose: bug fix
Description:
h5diff would incur a segmentation fault in Alpha clusters when
the percentage option is used.  This was due to memory aliagment
issue when the options variable is declared as a local variable
in main().

Moved the declaration to a global declaration.

Platforms tested:
"h5committested"
And tested at cluster QT of SNL.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@7418 dab4d1a6-ed17-0410-a064-d1ae371a2980