Purpose: bug fix.
Description:
It sometimes complained of wrong number of argument to test.
This happened when more than 1 tool are redeployed.
The problem is due to an unquoted variable expanded into
multiple strings, confusing the test command.
Fixed by putting quotes around it.

Platforms tested:
only tested in eirene since it is trivial.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@7809 dab4d1a6-ed17-0410-a064-d1ae371a2980