Purpose: h5diff new feature
Description:

added compare for attributes
a new options flag (-a) was added to the options structure. it is 0 by default (no compare )
the output of the compare is the same that for datasets, and all the other flags also apply for attributes
(the memory compare is done in the same function diff_array)
all the other requirements for compare of datasets (type, space) are identical too




Platforms tested:
linux
solaris 2.7
IRIX


Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@7813 dab4d1a6-ed17-0410-a064-d1ae371a2980