Purpose: Bug fix
Description:
MPIO driver did not detect partial read or write.  It assumed if read
completed, all bytes requested are read.  Therefore, if a dataset was
not completely written yet, the unread part will not be filled with
default fill value.

Solution:
Added the MPI_Get_elements call to find exactly how many bytes
are read (or written).  For read case, if not enough is read,
the rest is padded with zeros.  For write case, it is an error
if not able to write them all in first attempt. (It is not
always possible to repeat write for MPIO since file-view may
be changed, collective would hang, etc. So, we do just one attempt.)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@7972 dab4d1a6-ed17-0410-a064-d1ae371a2980