Purpose: h5repack change
Description:
changed the default size of chunks when defining a new dataset to apply a filter
this size is now equal to the current dimensions. previously , it was set to half this
value, and it was causing write calls to fail on some files (probably a bug)

Solution:

Platforms tested:
linux
solaris
AIX

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8056 dab4d1a6-ed17-0410-a064-d1ae371a2980