Purpose: Bug fix.
Description:
OSF1 machine at PSC barked at "make install" because an intended
comment was inserted in the middle of the column.  This is not kosher
because officially, make comments should be started with # at column 1.

Solution:
Moved the comment of installing h5fc as h5pfc to start at column 1.

Platforms tested:
Not h5committested since this is exposed by the PSC machine and Elena
has tested the fix there.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8073 dab4d1a6-ed17-0410-a064-d1ae371a2980