Purpose:     Bug fix.
Description:
    Revert the optimization to the MPI-I/O VFD from earlier this week that
eliminated calling MPI_File_set_size() when a file was truncated.  This doesn't
work because of the way we track the 'eoa' (end of allocated space) in the
file: we just advance the eoa value when space is allocated in the file and we
don't zero out the space in the file.  This causes a problem where we assume
that all "newly allocated" space in the file contains zero bytes, but if we
don't call MPI_File_set_size(), the file can contain the previous data (if the
file previously existed) and not zeros.
    Unfortunately, because we allow independent raw data parallel I/O, there
is no way to address this issue. :-(

Solution:
    Revert to previous method of retrieving the file's size, etc. from before
the optimization.

Platforms tested:
    FreeBSD 4.9 (sleipnir) w/parallel
    too minor to require h5committest.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8145 dab4d1a6-ed17-0410-a064-d1ae371a2980