Purpose:  Last step of check-in for Null dataspace
Description:  Mainly are header message changes for dataspace.  In last round
of check-in, a new header message for dataspace to created, which is not a good
way.  Now, there will be no new message for dataspace, but just add the type of
dataspace in the message while increment its version number.  Backward compatibility
is addressed.  The attribute design is modified accordingly.  Took out Null
dataspace test from tmisc.c and put it in th5s.c.


Platforms tested: h5committest


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8331 dab4d1a6-ed17-0410-a064-d1ae371a2980