Purpose: feature
Description:
H5check_version will print warning messages if the version does not
match or if the lib version string is not consistent, even when
$HDF5_DISABLE_VERSION_CHECK is set to 1.  This will mess up
tests that try to match output.

Changed it so that if $HDF5_DISABLE_VERSION_CHECK is set to 2, no
warning about version mismatch is printed.
The lib version string warning is suppressed if $HDF5_DISABLE_VERSION_CHECK
is set to any non-zero value.


Platforms tested:
h5committested.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8355 dab4d1a6-ed17-0410-a064-d1ae371a2980