Purpose: Handled SZIP without the encoder present.
Description:
It is now an error for a user to try to create, extend, or write to a
dataset without the encoder present in their SZIP library.
Added H5Zget_filter_info to provide users with a way to query HDF5
about the presence (or lack thereof) of the SZIP encoder.

Platforms tested:
  Windows
  Verbena
  Arabica
  Copper

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8753 dab4d1a6-ed17-0410-a064-d1ae371a2980