Purpose: Fixed a bug that broke daily tests. SZIP now always uses K13 compression.
Description:
When configured with the --srcdir option, dsets test would look for
a test file in the same directory as the test, instead of in the
directory that contains the source files.

SZIP never needs to use CHIP compression, so the option to set this flag
instead of K13 compression has been removed.



Platforms tested:
Copper


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8775 dab4d1a6-ed17-0410-a064-d1ae371a2980