Purpose: Bug fix and feature.
Description:
The setenv was done in runtest but it has the effect will linger
onto the next test.  So, if the first test sets $CXX to a certain
value, it lingers on the all following tests of the same host.
This is usually not desired.

Solution:
Move the actual setenv code to snapshot.  Runtest now just parses them
and pass along the setenv request to snapshot.

Platforms tested:
no h5comittest which does not really test the change.
Hand tested in eirene with Tg-NCSA.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8914 dab4d1a6-ed17-0410-a064-d1ae371a2980