Purpose: Improvement
Description:
Removed "-D__GNUC__" which was a patch for previous version.
It is no longer needed.
Also moved the debug flags "-w2 and -Wall" to DEBUG_CXXFLAGS
so that they are used for debug mode only.
Also adjust CXXFLAGS setting so that -std=c99 is added only
if user has not preset CXXFLAGS.  This allows users the option
of not using it.

Platforms tested:
Tested in tg-NCSA only since that is the only ia64 platform
around.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@8934 dab4d1a6-ed17-0410-a064-d1ae371a2980