Purpose: bug fix.
Description:
The test routines only print error messages but not all of them
return number of errors detected back to the main routine which
always exit with a 0 status.  Thus make or shell commands could
not detect there were errors.


Solution:
Changed the test routines to return appropriate number of
errors to main routine which in turn exit with the appropriate
exit code if errors found.

Platforms tested:
Tested in Sol and eirene (pp).

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9130 dab4d1a6-ed17-0410-a064-d1ae371a2980