Purpose: Fix parallel bug reported by Thomas Guignon, in which different processes became confused as to whether they were doing collective or individual I/O.
Description:
When one process had a point selection, and another didn't, the
first concluded that the I/O was independant, while the second
presumed that it was collective.  A hang resulted.


Solution:
Get all processes involved in an I/O to compare notes.  If all
agree that the I/O is collective, they proceed with same.  If
any think the I/O should be independant, all use independant
I/O.

Note that this is an interrim fix -- the correct solution is to
support collective I/O on point selections.  This will take a
while.


Platforms tested:
copper

h5committested

Eirene (serial and parallel)

In the parallel test on Eirene, I encountered a bug in h5repacktst.
However the problem vanished on recompile.  Since I couldn't reproduce
it elsewhere, I went ahead with the checkin.

Given my druthers, I would have liked to study the code more
carefully before this check-in.  However, there is some time pressure.

The new code implementing the consensus check must not be executed
unless MPI is initialized, and there is a communicator associated
with the file.  I think my guards against this case are adequate,
but if we run into a hang or an illegal instruction error, this
change should be suspect.


Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9178 dab4d1a6-ed17-0410-a064-d1ae371a2980