Purpose: Bug fix.
Description:
There was a historical hack to prevent runtest in multiple
hosts to run cvs update while they were launched by the daily test.
But a more proper fix has been implemented to prevent that but
this hack was not removed.  It ended up enforcing one CVS update
per day.  But if someone tested runtest during the day, it would
prevent the evening daily from doing update.  That is not good.

Solution:
Removed the hack which is not need any more.

Platforms tested:
NO h5committest since it does test runtest.  Tested it by
hand in eirene and heping.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9339 dab4d1a6-ed17-0410-a064-d1ae371a2980