Purpose: Changed functionality of H5Pget_data_transform
Description:
H5Pget_data_transform no longer allocates memory for the transform string...it
is the user's responsbility to do so.

Solution:
Made H5Pget_data_transform be more in line with other functions that do
similiar things:

User now has to allocate memory for the string themselves and specify how much
of the string should be copied into the buffer.  There is also support for
querying the length of the transform string.

Platforms tested:
sol + eirene + copper

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9461 dab4d1a6-ed17-0410-a064-d1ae371a2980