Purpose: Bug fix.
Description:
MIPSpro Compiler 7.4.x starts to support C99 features
but stdint.h has a "guard" in it that will #error if
cc is invoked without asking for C99 standard.  But it
does not result with a non-zero compiler exit code, thus
configure thought it is okay to use header file stdint.h.
This causes problems later.

Solution:
Default to use the C99 compiler if available.

Platforms tested:
Tested in Sandia tesla machine which is a new SGI machine.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9660 dab4d1a6-ed17-0410-a064-d1ae371a2980