Purpose:
Fix a bug in the cache caused by a slightly over active sanity check.


Description:

When the #define H5C_DO_SANITY_CHECKS is TRUE, the macro
H5C__DLL_PRE_INSERT_SC is executed prior to inserting an
entry in the LRU list.  The macro performs a variety of
sanity checks, and flags an error if any of the checks
fail.

Prior to this update, the macro used to check to see if
the target list had length 1 and size <= 0.  The new
epoch marker entries used in the age out cache size
reduction algorithm have size zero -- making it possible
for this sanity check to fail incorrectly.

Note that cache sanity checks are disabled in the CVS
version of the code, so this bug and bug fix should be
invisible unless you are working with the cache, and
turn the sanity checks on.


Solution:

Removed the offending clause in H5C__DLL_PRE_INSERT_SC.  Since
the size used in the macro is typically a size_t and thus
cannot have negative value, there was no point in changing
it to "size < 0".


Platforms tested:

Ran a serial test on heping.

Under the circumstances, I didn't feel the need for further
testing.


Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9700 dab4d1a6-ed17-0410-a064-d1ae371a2980