Purpose:
Reduce run time of daily tests.


Description:

cache, the test program for the metadata cache has been taking a while
to execute.


Solution:

As a short term "fix", I have commented out all but one of the long
running test functions.  Of course that means that we aren't running
these tests at present.  I'm not sure that this is a good idea.


Platforms tested:

Serial on Heping.


Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@9790 dab4d1a6-ed17-0410-a064-d1ae371a2980