Purpose: Configure feature
Description:
Changed bin/reconfigure.sh script to use autotools in AFS.

Solution:
Previously, the only machine with the correct versions of autoconf and
automake was heping.  Now both tools are installed in AFS, so
in theory any hdf machine can be used to run the reconfigure script.

Platforms tested:
heping, eirene, verbena.
On sleipnir and arabica the autotools were unable to find a version
of m4 that was new enough for them.

Misc. update:


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10005 dab4d1a6-ed17-0410-a064-d1ae371a2980