Purpose: Bug fix
Description:
HDF5's libtool includes a hack to ensure that libraries are found
on IRIX.  This hack did not have the correct path for the hl-fortran
tests to find the main library.

Solution:
Extend the hack.  This may call for a better long-term solution...

Platforms tested:
modi4 (hack only takes effect on IRIX)


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5@10048 dab4d1a6-ed17-0410-a064-d1ae371a2980